Added the app logo Added the onboarding screens Modified the xml files
