+++ app/build.gradle.kts
@@ -8,18 +8,16 @@ plugins {
 }
 
 android {
-    namespace = "com.example.notetakingapp"
+    namespace = "com.note.notetakingapp"
     compileSdk = 34
-
     defaultConfig {
-        applicationId = "com.example.notetakingapp"
+        applicationId = "com.note.notetakingapp"
         minSdk = 27
         targetSdk = 34
         versionCode = 1
         versionName = "1.0"
         testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
     }
-
     buildTypes {
         release {
             isMinifyEnabled = false
@@ -28,17 +26,15 @@ android {
                 "proguard-rules.pro"
             )
         }
-    }
 
+    }
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
     }
-
     kotlinOptions {
         jvmTarget = "1.8"
     }
-
     buildFeatures {
         dataBinding = true
         viewBinding = true
@@ -46,9 +42,7 @@ android {
     kapt{
         correctErrorTypes=true
     }
-
     }
-
 dependencies {
     implementation("androidx.core:core-ktx:1.13.1")
     implementation("org.jetbrains.kotlinx:kotlinx-coroutines-android:1.8.1") // Updated version
@@ -75,4 +69,5 @@ dependencies {
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.2.1")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.6.1")
+
 }
