--- app/src/main/java/com/example/notetakingapp/repository/NotesRepository.kt
@@ -1,41 +1,0 @@
-package com.example.notetakingapp.repository
-
-import com.example.notetakingapp.room.Note
-import com.example.notetakingapp.room.NotesDao
-import com.example.notetakingapp.room.Priority
-
-class NotesRepository(private val db: NotesDao) {
-
-    val allnotes = db.getAllNotes()
-
-    suspend fun insertNote(notes: Note) {
-        return db.insertNote(notes)
-    }
-
-    suspend fun update(notes: Note) {
-        return db.updateNote(notes)
-    }
-
-    suspend fun delete(notes: Note) {
-        return db.deleteNote(notes)
-    }
-
-    suspend fun deleteNotesById(id: Int) {
-        return db.deleteNotesById(id)
-    }
-
-    fun searchNote(query: String?) = db.searchNote(query)
-
-    suspend fun updateIsPinnedColumn(noteIds: List<Int>, isPinned: Int) {
-        db.updateIsPinnedColumn(noteIds, isPinned)
-    }
-
-    suspend fun updatePriorityColumnInDB(
-        noteIds: List<Int>,
-        priority: Priority,
-        isHighPriorityVisible: Int
-    ) {
-        db.updatePriorityColumnInDB(noteIds, priority, isHighPriorityVisible)
-    }
-
-}
