--- app/src/main/java/com/example/notetakingapp/util/Constants.kt
@@ -1,6 +1,0 @@
-package com.example.notetakingapp.util
-
-object Constants {
-    const val PREF_NAME = "app_prefs"
-    const val PREF_KEY_ONBOARDING = "onboarding_complete"
-}
