--- app/src/main/java/com/example/notetakingapp/view/ViewPageAdapter.kt
@@ -1,22 +1,0 @@
-package com.example.notetakingapp.view
-
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentManager
-import androidx.fragment.app.ListFragment
-import androidx.lifecycle.Lifecycle
-import androidx.viewpager2.adapter.FragmentStateAdapter
-
-class ViewPageAdapter (
-    private val fragmentList: ArrayList<Fragment>,
-    fragmentManager: FragmentManager,
-    lifecycle: Lifecycle
-):FragmentStateAdapter(fragmentManager,lifecycle) {
-
-    override fun getItemCount(): Int {
-        return fragmentList.size
-    }
-
-    override fun createFragment(position: Int): Fragment {
-        return fragmentList[position]
-    }
-}
