--- app/src/main/java/com/example/notetakingapp/view/screens/SecondScreen.kt
@@ -1,22 +1,0 @@
-package com.example.notetakingapp.view.screens
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import com.example.notetakingapp.R
-
-
-class SecondScreen : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_secondscreen, container, false)
-        return view
-    }
-}
-
