--- app/src/main/java/com/example/notetakingapp/view/screens/ThirdScreen.kt
@@ -1,23 +1,0 @@
-package com.example.notetakingapp.view.screens
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import com.example.notetakingapp.R
-
-
-class ThirdScreen : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        val view = inflater.inflate(R.layout.fragment_thirdscreen, container, false)
-        return view
-    }
-}
-
-
