--- app/src/main/java/com/example/notetakingapp/viewmodel/ViewModelFactory.kt
@@ -1,23 +1,0 @@
-package com.example.notetakingapp.viewmodel
-
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.ViewModelProvider
-import com.example.notetakingapp.repository.NotesRepository
-
-class ViewModelFactory (private val repo: NotesRepository) : ViewModelProvider.Factory{
-    // this is a class which is used to instantiate and
-    override fun <T : ViewModel> create(modelClass: Class<T>): T {
-        if(modelClass.isAssignableFrom(NoteViewModel::class.java)){
-            return NoteViewModel(repo) as  T
-            // here we are creating a contact viewmodel in  create fun
-        }
-        throw IllegalArgumentException("Unknown view model class")
-
-
-    }
-
-}
-
-
-
-
