+++ app/src/main/java/com/note/notetakingapp/util/Constants.kt
@@ -1,0 +1,6 @@
+package com.note.notetakingapp.util
+
+object Constants {
+    const val PREF_NAME = "app_prefs"
+    const val PREF_KEY_ONBOARDING = "onboarding_complete"
+}
