+++ app/src/main/java/com/note/notetakingapp/view/screens/FirstScreen.kt
@@ -1,0 +1,20 @@
+package com.note.notetakingapp.view.screens
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.note.notetakingapp.R
+
+class FirstScreen : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        val view = inflater.inflate(R.layout.fragment_firstscreen, container, false)
+        return view
+    }
+}
