+++ app/src/main/java/com/note/notetakingapp/view/screens/SecondScreen.kt
@@ -1,0 +1,22 @@
+package com.note.notetakingapp.view.screens
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import com.note.notetakingapp.R
+
+
+class SecondScreen : Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        val view = inflater.inflate(R.layout.fragment_secondscreen, container, false)
+        return view
+    }
+}
+
