+++ app/src/main/res/layout/fragment_secondaryfragment.xml
@@ -6,7 +6,7 @@
     <data>
         <variable
             name="notes"
-            type="com.example.notetakingapp.viewmodel.NoteViewModel" />
+            type="com.note.notetakingapp.viewmodel.NoteViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
